Raise the hull vertex cap to 64
Slice 10-2, the deep dive's parity number. The caps move by
Euler's ledger for V = 64: faces to 124 (2V - 4), half edges and
face indices to 372 (6V - 12), and the QuickHull input cloud cap
to 256 points. Half-edge twin and next indices outgrow a byte at
this scale, so they widen to uint16 (vertex and face indices stay
bytes), and faceVertStart widens with them; the three narrowing
writers in the builder and the box-hull path follow. The QH
work-pool caps derive from the constants and re-derive
themselves; the SAT and clipping stack buffers grow from the same
defines.

Hull geometry is create-time state the hash covers by INDEX, so
nothing moved: golden and every bench pin stand bit-for-bit. Only
the snapshot block grew (1820 to 5808 bytes per hull slot, format
v38), which the vehicle and soft-body suites felt as too-small
snapshot buffers: grown, and the ledger notes the honest cause.

Suite growth: test_quickhull builds a 128-point golden-angle
sphere cloud into a hull RICHER than the old 24-vertex ceiling,
bounded by the new caps, twin-paired at scale, and refuses 257
input points; test_compound drops a 96-point rock that rests near
its radius and sleeps (with rolling resistance, honoring the 6-3
lesson instead of relearning it) with bit-identical twins.

+++ src/hull.c
@@ -48,7 +48,7 @@ void m3BuildBoxHull(m3HullData* out, m3Vec3 halfExtents)
     for (int32_t f = 0; f < 6; ++f)
     {
         out->faceVertCounts[f] = 4;
-        out->faceVertStart[f] = (uint8_t)(4 * f);
+        out->faceVertStart[f] = (uint16_t)(4 * f);
         for (int32_t k = 0; k < 4; ++k)
         {
             out->faceIndices[4 * f + k] = loops[f][k];
@@ -113,7 +113,7 @@ void m3HullBuildHalfEdges(m3HullData* hull)
                 dirCount += 1;
             }
             hull->edges[slot].origin = (uint8_t)a;
-            hull->edges[slot].twin = (uint8_t)(slot ^ 1);
+            hull->edges[slot].twin = (uint16_t)(slot ^ 1);
             hull->edges[slot].face = (uint8_t)f;
             faceEdgeSlot[start + k] = slot;
         }
@@ -125,7 +125,7 @@ void m3HullBuildHalfEdges(m3HullData* hull)
         int32_t start = hull->faceVertStart[f];
         for (int32_t k = 0; k < n; ++k)
         {
-            hull->edges[faceEdgeSlot[start + k]].next = (uint8_t)faceEdgeSlot[start + (k + 1) % n];
+            hull->edges[faceEdgeSlot[start + k]].next = (uint16_t)faceEdgeSlot[start + (k + 1) % n];
         }
     }
     hull->edgeCount = count;
